Refactor Send Message and add functionality
I made a couple of structural changes to the code just in the send_text
function, but can be used in further functionality for new modules. The
basic idea is to encapsulate the Blower API call in one function and
then validate the use through the login. This prevents you from having
to “login wall” each page and the login only becomes relevant if a
person actually wants to send a message